Make StreamMixin a part of core client
