+++ contrib/ustreamer/jumpstarter_driver_ustreamer/client.py
@@ -1,12 +1,11 @@
 from base64 import b64decode
 
 from jumpstarter.client import DriverClient
-from jumpstarter.client.mixins import StreamMixin
 
 from .common import UStreamerState
 
 
-class UStreamerClient(DriverClient, StreamMixin):
+class UStreamerClient(DriverClient):
     def state(self):
         """
         Get state of ustreamer service
