+++ docs/source/reference.md
@@ -52,7 +52,6 @@ from contextlib import asynccontextmanager
 from collections.abc import Generator
 from jumpstarter.driver import Driver, export, exportstream
 from jumpstarter.client import DriverClient
-from jumpstarter.client.mixins import StreamMixin
 from jumpstarter.common.utils import serve
 
 class ExampleDriver(Driver):
@@ -83,7 +82,7 @@ class ExampleDriver(Driver):
         async with await connect_tcp(remote_host="example.com", remote_port=80) as stream:
             yield stream
 
-class ExampleClient(DriverClient, StreamMixin):
+class ExampleClient(DriverClient):
     # client methods are sync
     def echo(self, message) -> str:
         return self.call("echo", message)
