+++ jumpstarter/drivers/network/client.py
@@ -1,6 +1,5 @@
 from jumpstarter.client import DriverClient
-from jumpstarter.client.mixins import StreamMixin
 
 
-class NetworkClient(DriverClient, StreamMixin):
+class NetworkClient(DriverClient):
     pass
