+++ jumpstarter/drivers/pyserial/driver_test.py
@@ -4,6 +4,6 @@ from jumpstarter.drivers.pyserial.driver import PySerial
 
 def test_drivers_pyserial():
     with serve(PySerial(url="loop://")) as client:
-        with client.connect() as stream:
+        with client.stream() as stream:
             stream.send(b"hello")
             assert "hello".startswith(stream.receive().decode("utf-8"))
