+++ jumpstarter/streams/__init__.py
@@ -1,5 +1,6 @@
+from .blocking import BlockingStream
 from .common import create_memory_stream, forward_stream
 from .router import RouterStream
 from .websocket import WebsocketServerStream
 
-__all__ = ["create_memory_stream", "forward_stream", "RouterStream", "WebsocketServerStream"]
+__all__ = ["create_memory_stream", "forward_stream", "BlockingStream", "RouterStream", "WebsocketServerStream"]
